+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSVillagesConfig.java
@@ -15,12 +15,25 @@ public class RSVillagesConfig
 		public ConfigValueListener<Integer> swampVillageMaxChunkDistance;
 		public ConfigValueListener<Integer> mountainsVillageMaxChunkDistance;
 		public ConfigValueListener<Integer> giantTaigaVillageMaxChunkDistance;
+		public ConfigValueListener<Integer> oakVillageMaxChunkDistance;
 		public ConfigValueListener<Integer> crimsonVillageMaxChunkDistance;
 		public ConfigValueListener<Integer> warpedVillageMaxChunkDistance;
-		public ConfigValueListener<Integer> villageOakMaxChunkDistance;
+
+		public ConfigValueListener<Integer> badlandsVillageSize;
+		public ConfigValueListener<Integer> birchVillageSize;
+		public ConfigValueListener<Integer> darkForestVillageSize;
+		public ConfigValueListener<Integer> jungleVillageSize;
+		public ConfigValueListener<Integer> swampVillageSize;
+		public ConfigValueListener<Integer> mountainsVillageSize;
+		public ConfigValueListener<Integer> giantTaigaVillageSize;
+		public ConfigValueListener<Integer> oakVillageSize;
+		public ConfigValueListener<Integer> crimsonVillageSize;
+		public ConfigValueListener<Integer> warpedVillageSize;
 
 		public RSVillagesConfigValues(ForgeConfigSpec.Builder builder, ConfigHelper.Subscriber subscriber)
 		{
+			builder.push("Spawnrates");
+
 			badlandsVillageMaxChunkDistance = subscriber.subscribe(builder
 				.comment("\n How rare are Badlands Villages in Badland biomes.",
 					" 1 for spawning in most chunks and 1001 for no spawn.")
@@ -63,11 +76,11 @@ public class RSVillagesConfig
 				.translation("repurposedstructures.config.villages.gianttaigavillagemaxchunkdistance")
 				.defineInRange("giantTaigaVillageMaxChunkDistance", 47, 1, 1001));
 
-			villageOakMaxChunkDistance = subscriber.subscribe(builder
+			oakVillageMaxChunkDistance = subscriber.subscribe(builder
 					.comment("\n How rare are Oak Villages in forest category",
 							" biomes that are not birch or dark forest.")
 					.translation("repurposedstructures.config.village.villageoakmaxchunkdistance")
-					.defineInRange("villageOakMaxChunkDistance", 47, 1, 1001));
+					.defineInRange("oakVillageMaxChunkDistance", 47, 1, 1001));
 
 			crimsonVillageMaxChunkDistance = subscriber.subscribe(builder
 				.comment("\n How rare are Crimson Village in Crimson Forest biomes.",
@@ -80,6 +93,62 @@ public class RSVillagesConfig
 						 " 1 for spawning in most chunks and 1001 for none.")
 				.translation("repurposedstructures.config.villages.warpedvillagemaxchunkdistance")
 				.defineInRange("warpedVillageMaxChunkDistance", 30, 1, 1001));
+
+			builder.pop();
+
+			builder.push("Size");
+
+			badlandsVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.badlandsvillagesize")
+					.defineInRange("badlandsVillageSize", 10, 1, 30));
+
+			birchVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.birchvillagesize")
+					.defineInRange("birchVillageSize", 6, 1, 30));
+
+			darkForestVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.darkforestvillagesize")
+					.defineInRange("darkForestVillageSize", 6, 1, 30));
+
+			jungleVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.junglevillagesize")
+					.defineInRange("jungleVillageSize", 8, 1, 30));
+
+			swampVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.swampvillagesize")
+					.defineInRange("swampVillageSize", 6, 1, 30));
+
+			mountainsVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.mountainsvillagesize")
+					.defineInRange("mountainsVillageSize", 6, 1, 30));
+
+			giantTaigaVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.gianttaigavillagesize")
+					.defineInRange("giantTaigaVillageSize", 6, 1, 30));
+
+			oakVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.oakvillagesize")
+					.defineInRange("oakVillageSize", 6, 1, 30));
+
+			crimsonVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.crimsonvillagesize")
+					.defineInRange("crimsonVillageSize", 6, 1, 30));
+
+			warpedVillageSize = subscriber.subscribe(builder
+					.comment("\n Size of the Village. This is how many pieces long a path can be from the start piece.")
+					.translation("repurposedstructures.config.warpedvillagesize")
+					.defineInRange("warpedVillageSize", 6, 1, 30));
+
+			builder.pop();
 		}
 	}
 }